Remove add action as that is now covered by pull-list.py
Remove option to skip the pull, as without the add action that is all
this script now does.

Allow script to be run sharded.  Experimentation with threading didn't
really show much of a performance improvement.  New behaviour allows you
to run a job for just a subset of volumes.  You can then run a number of
these jobs independently.

I am running this using a crontab to run it once an hour, with 24
workers and processing shard $(date +%H) (i.e. during a 24 hour period
all volumes will be updated).  This means that my pull database is
locked for approx 1 minute per hour rather than 16 minutes all in one
go.