Alter threshold for changes to something with more of a mathematical footing.
In a normal distribution 50% of values are within 0.675 standard
deviations of the mean.  The median is the value which has 50% of values
above and 50% of values below.  For both of these to hold true the
median needs to be within 0.675 standard deviations of the mean.

Improve logging.

Alter stream identification pattern to try and avoid embedded '+'
characters in titles causing confusion.